nlp_assignment/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
newsapi-python==0.2.7
requests==2.30.0
urllib3==2.0.2
